Support for pure functions with n parameters.
functions are now picked up by the tokenizer and evaluator. the idea is
that the contents of their parentheses may reference a different
namespace, so the evaluator isn't meant to look inside the
parentheses (it currently does though, I think).

this can be tested by running
`exp('invert(B,b) + a + b = c').display()` in a python environment with axioms_2
loaded. `invert` is currently a dummy function, and is just there for demonstration.