support for arbitrary functions in command line
arbitrary functions (such as invert) can be called in the command line
while setting expressions. when an expression is created, it can be
evaluated for arbitrary commands. commands which are recognized are
expanded. Their results are spliced into the current expression tree.

also added the 'eval' command, which does this in the command line
window.